+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - [feat: add metadata transfer for http header via spring.cloud.tencent.metadata.headers.](https://github.com/Tencent/spring-cloud-tencent/pull/1175)
 - [fix:remove bcprov-jdk15on dependency.](https://github.com/Tencent/spring-cloud-tencent/pull/1179)
 - [feat:support configuration encryption.](https://github.com/Tencent/spring-cloud-tencent/pull/1183)
+- [feat:optimize examples.](https://github.com/Tencent/spring-cloud-tencent/pull/1187)